+++ src/senaite/storage/setuphandlers.py
@@ -20,8 +20,6 @@
 
 from Acquisition import aq_base
 from bika.lims import api
-from plone import api as ploneapi
-from plone.app.dexterity.behaviors.exclfromnav import IExcludeFromNavigation
 from Products.CMFCore.permissions import ModifyPortalContent
 from Products.DCWorkflow.Guard import Guard
 from senaite.core import permissions
@@ -35,7 +33,6 @@ from senaite.storage.catalog import STORAGE_CATALOG
 from senaite.storage.catalog import StorageCatalog
 from senaite.storage.config import PRODUCT_NAME
 from senaite.storage.config import PROFILE_ID
-
 
 SITE_STRUCTURE = [
     # Tuples of (portal_type, obj_id, obj_title, parent_path, display_type)
@@ -429,21 +426,27 @@ def setup_site_structure(portal):
 
 
 def display_in_nav(obj):
-    """Makes an object to be displayed in the navigation bar
+    """Makes an object and/or objects from the given portal type to be
+    displayed in the navigation bar
     """
-    # Display in navigation
-    registry_id = "plone.displayed_types"
     portal_type = api.get_portal_type(obj)
-    to_display = ploneapi.portal.get_registry_record(registry_id, default=())
-    if portal_type not in to_display:
-        to_display = to_display + (portal_type, )
-        ploneapi.portal.set_registry_record(registry_id, to_display)
 
-    nav_exclude = IExcludeFromNavigation(obj, None)
-    if nav_exclude:
-        nav_exclude.exclude_from_nav = False
-        obj.reindexObject(idxs=["exclude_from_nav"])
+    # remove from senaite setup's sidebar_skip_types
+    setup = api.get_senaite_setup()
+    skip = setup.getSidebarSkipTypes()
+    if skip and portal_type in skip:
+        skip = tuple(pt for pt in skip if pt != portal_type)
+        setup.setSidebarSkipTypes(skip)
 
+    # if a root folder, add to senaite setup's sidebar_folders
+    setup = api.get_senaite_setup()
+    portal = api.get_portal()
+    if api.get_parent(obj) == portal:
+        obj_id = api.get_id(obj)
+        folders = setup.getSidebarFolders()
+        if obj_id not in folders:
+            folders += (obj_id, )
+            setup.setSidebarFolders(folders)
 
 def reindex_storage_structure(portal):
     """Reindex storage structure